renaming of directories
