+++ fesca/computing_node/src/receiving_shares/storage.rs
